+++ drone_new/src/rl_lib.h
@@ -14,8 +14,8 @@ struct EulerAngles {
 
 EulerAngles ToEulerAngles(Quaternion q);
 void quat2eular_show(const Eigen::Quaterniond &quatf);
-Eigen::Vector4f quatMultiplication(const Eigen::Vector4f &q, const Eigen::Vector4f &p);
-Eigen::Vector4f rot2Quaternion(const Eigen::Matrix3d &R);
-Eigen::Matrix3d quat2RotMatrix(const Eigen::Vector4f &q);
+Eigen::Vector4d quatMultiplication(const Eigen::Vector4d &q, const Eigen::Vector4d &p);
+Eigen::Vector4d rot2Quaternion(const Eigen::Matrix3d &R);
+Eigen::Matrix3d quat2RotMatrix(const Eigen::Vector4d &q);
 
 #endif
